Change myList to PFQueryTableView
Make myList saved in cache, so that it will not request to server
everytime user opens myList.
Also implement pull-down refresh.
Load S3 URL one by one when user tap one of the video.
Need to save the URL or the video itself in client device so that he
does not have to call to AWS again.