Modify mainView & Add HUD to other Views
Make mainView saved in cache, so that it will not request to server
every time user opens News Feed.
Also implement pull-down refresh.
Load S3 URL one by one when user taps one of the video.
Add videoType, userName, status to the table cell view.
Add HUD to other Views in order to make the loading smoother.
Need to save the URL of the video itself in client device so that he
does not have to call to AWS again.