Add userId, username and targetVideo's Id into videoURL
To make it easy to deal with aws s3.
We think of in the future, users may want to retake answer for a
question because say they’ve improved presentation skill.
But they would like to look back to their previous answer. So we want
to save the aws s3 link in the form like:
userId_username_targetObjectId_uploadDate.mov